Refactor and add text measuring
The refactoring involves tightening the buffer looping in a way such
that both measuring and printing use the exact same code. Because
measuring is in fact just printing without actually drawing, this makes
a lot of sense, and let us avoid having to maintain the same loop
twice- which is an inherently error-prone way to go about it.