Refactor and add additional print functions
The new `printstr` variations is intended to separate glyph drawing
from text drawing; of course, behind the scenes they are the same-
however, for user code it offers a neat distinction between what is
“game objects” and what is meant as readable text.

You often don’t want to specify bounds and wrapping for stuff that is
meant to be drawn as-is. You do want it for long pieces of text,
however.