Add fill functions for conveniently coloring an area
This is essentially an extension to the existing drawing functions,
which uses a trick to provide rect drawing capabilities. Behind the
scenes it just stretches the full-block glyph to fit the desired area.

The fill functions are not core to the project and will likely be moved
or pulled out as an extension API. However, previously, a typical
pattern was to make a loop and print the full-block glyph several times
until an area was covered. This was not very efficient and needed a
solution.

It could obviously be done cheaper and faster by implementing this as
dedicated renderer functionality, but this solution is preferred (at
least for now) as it utilizes existing APIs.