Fix CP437 example
This example was critically flawed. It incorrectly assumed strings
passed to `term_print` would be copied and stored internally; this is
not the case.

The result was an unpredictable output of character glyphs, as the same
4-byte large array would be written to over and over.

Additionally, the string buffer was allocated in function scope, which
was also a source of unpredictability. Since that buffer would be
released upon exiting the draw function, the pointer passed to the
print command could point to anything (and the draw function would
*always* exit before any print commands were actually flushed).

I'm a bit on the fence about whether or not strings passed *should*
actually be copied internally; this would make it much simpler to just
fire off print commands from anywhere, but it would also take up more
responsibility and memory from something that kinda actually should be
up to the client to handle...