Refactor and improve text buffering performance
This commit enables `buffer` to store a decoded version of a copied
string and adds pointer comparison to determine whether the string has
changed, to avoid doing the same work over and over.

This is a pretty good performance gain for strings that are repeatedly
printed, because it works as a cache.

However, it does have a significant flaw: it determines whether to use
cached content by comparing a pointer reference- but, a pointer
reference does not necessarily change when the content it points to
does. So in a scenario where a program is printing only a single
string, but that string changes its content continuously, the program
would only ever show the first string it printed. Blazingly fast, mind
you- but not correct.