Reduce function call overhead by caching common values
I feel like this particular area could use some refactoring as I'm not
too fond of this solution.

For example, do we really need a "font object"? The whole idea is that
we've embedded a specific font, and thus the size and information won't
be changing, so why not just have #defines somewhere?