Revert caching optimization
Refer to 3443763005a6b176923a4225f4f8f9f80fd9af27

This optimization makes a (very) negligible difference, at the cost of having more obscure code. In practice it meant keeping track of the same stuff twice, having to make sure it was always in sync (specifically the viewport) which is just problems waiting to happen.